LayoutTests/imported/w3c/web-platform-tests/navigation-api/precommit-handler/precommitHandler-redirect-push-changed-to-replace-expected.txt


FAIL precommitHandler redirect() push changed to replace assert_equals: expected "#redirect1" but got "#push"
